+++ src/chat/ai/google_gemini.py
@@ -3,30 +3,29 @@ from typing import Optional, Dict, Any
 
 import litellm
 
-from chat.llm_provider import LLMProvider
-from chat.conversation import Conversation, MessageType
-from chat.logging_util import logger
+from chat.ai.llm_provider import LLMProvider
+from chat.conversation.conversation import Conversation, MessageType
+from chat.util.logging_util import logger
 
 
-class AnthropicProvider(LLMProvider):
-    """Integration with Anthropic Claude models using LiteLLM."""
+class GoogleGeminiProvider(LLMProvider):
+    """Integration with Google Gemini models using LiteLLM."""
 
-    def __init__(self, api_key: Optional[str] = None, model: str = "claude-3-7-sonnet-latest"):
-        self.api_key = api_key or os.getenv("ANTHROPIC_API_KEY")
+    def __init__(self, api_key: Optional[str] = None, model: str = "gemini-2-flash"):
+        self.api_key = api_key or os.getenv("GOOGLE_API_KEY")
         if not self.api_key:
-            raise ValueError("Anthropic API key is required. Set it in .env or as an environment variable.")
+            raise ValueError("Google API key is required. Set it in .env or as an environment variable.")
 
+        # Use LiteLLM's model naming convention for Gemini
         self.model = model
+        if not self.model.startswith("gemini/"):
+            self.model = f"gemini/{model}"
 
-        # Use LiteLLM's model naming convention for Anthropic
-        if not self.model.startswith("anthropic/"):
-            self.model = f"anthropic/{model}"
-
-        os.environ["ANTHROPIC_API_KEY"] = self.api_key
+        os.environ["GEMINI_API_KEY"] = self.api_key
 
         try:
             self.client = litellm
-            logger.info(f"AnthropicProvider initialized with model: {self.model}")
+            logger.info(f"GoogleGeminiProvider initialized with model: {self.model}")
         except ImportError:
             logger.error("litellm package not installed. Please install it (e.g., pip install litellm)")
             raise
@@ -38,21 +37,19 @@ class AnthropicProvider(LLMProvider):
             options: Optional[Dict[str, Any]] = None,
             conversation: Optional[Conversation] = None
     ) -> str:
-        """Generate a completion from Claude using LiteLLM."""
+        """Generate a completion from Gemini using LiteLLM."""
         options = options or {}
         max_tokens = options.get("max_tokens", 64000)
         temperature = options.get("temperature", 0.7)
         system_prompt = options.get("system_prompt",
                                     "You are a helpful assistant specializing in providing accurate and relevant information.")
 
-        logger.info(f"Sending request to Claude with model: {self.model}")
+        logger.info(f"Sending request to Gemini with model: {self.model}")
 
         try:
             # Determine message format based on conversation history
             if conversation and conversation.messages:
                 # Add the new prompt if it's not already the last user message
-                # This prevents duplicating the prompt if it was already added to the conversation
-                # by a previous call to this method
                 messages = conversation.to_llm_messages()
 
                 # Ensure a system message exists at the beginning
@@ -80,7 +77,7 @@ class AnthropicProvider(LLMProvider):
 
             output = response.choices[0].message.content
             reason = response.choices[0].finish_reason or "unknown"
-            logger.info(f"Received response from Claude. Finish reason: {reason}, Output length: {len(output or '')}")
+            logger.info(f"Received response from Gemini. Finish reason: {reason}, Output length: {len(output or '')}")
 
             # Update conversation if provided
             if conversation and output:
@@ -89,7 +86,7 @@ class AnthropicProvider(LLMProvider):
             return output or ""
 
         except Exception as e:
-            logger.error(f"Error generating completion from Anthropic via LiteLLM: {e}", exc_info=True)
+            logger.error(f"Error generating completion from Google Gemini via LiteLLM: {e}", exc_info=True)
             raise
 
     async def generate_json(
@@ -99,5 +96,5 @@ class AnthropicProvider(LLMProvider):
             options: Optional[Dict[str, Any]] = None,
             conversation: Optional[Conversation] = None
     ) -> Dict[str, Any]:
-        """Generate a JSON response from Claude using the parent class implementation."""
+        """Generate a JSON response from Gemini using the parent class implementation."""
         return await super().generate_json(prompt, schema, options, conversation)
