+++ src/chat/ai/open_ai.py
@@ -1,11 +1,11 @@
 import os
-from typing import Optional, Dict, Any, Literal, List
+from typing import Optional, Dict, Any, Literal
 
 import litellm
 
-from chat.llm_provider import LLMProvider
-from chat.conversation import Conversation, MessageType
-from chat.logging_util import logger
+from chat.ai.llm_provider import LLMProvider
+from chat.conversation.conversation import Conversation, MessageType
+from chat.util.logging_util import logger
 
 
 class OpenAIProvider(LLMProvider):
